+++ addon/objects/uml-primitives/fd-uml-object-in-state.js
@@ -78,8 +78,8 @@ export let ObjectInState = ActiveState.define('flexberry.uml.ObjectInState', {
   attrs: {
     '.flexberry-uml-header-rect': { 
       'rx': 0,
-      'ry': 0
-    }
+      'ry': 0,
+    },
   },
 });